+++ tests/column_view/align.c
@@ -8,7 +8,7 @@
 #include "test.h"
 
 static void column_line_print(const void *data, int column_id, const char buf[],
-		size_t offset, AlignType align);
+		int offset, AlignType align);
 static void column1_func(void *data, size_t buf_len, char buf[],
 		const format_info_t *info);
 static void column2_func(void *data, size_t buf_len, char buf[],
@@ -33,8 +33,8 @@ TEARDOWN()
 }
 
 static void
-column_line_print(const void *data, int column_id, const char buf[],
-		size_t offset, AlignType align)
+column_line_print(const void *data, int column_id, const char buf[], int offset,
+		AlignType align)
 {
 	if(column_id == COL1_ID)
 	{
